+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 2018.7.16
 ========
 
+- Resolved #922: Change to a new editor.
+- Fixed #877: Fix uploading images in editor via Firefox.
 - Fixes #399: Filter image selection for images only via backend.
+
 
 2018.7.9
 ========